@@ -20,9 +20,8 @@ import (
 
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
-	"github.com/prometheus/common/route"
-
 	dto "github.com/prometheus/client_model/go"
+	"github.com/prometheus/common/route"
 
 	"github.com/prometheus/pushgateway/handler"
 	"github.com/prometheus/pushgateway/storage"
@@ -168,7 +167,7 @@ type response struct {
 
 func (api *API) respond(w http.ResponseWriter, data interface{}) {
 	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(200)
+	w.WriteHeader(http.StatusOK)
 
 	b, err := json.Marshal(&response{
 		Status: statusSuccess,
